Empty key as empty list instead of None
Without this -

If you pass no keys parameter to Droplet.__init__(), it still tries to
iterate upon registered keys to match their name. But as its default
value is None, error is raised.

With this -

keys is set to empty list instead of None, so that the iteration doesnt
raise error.